add frontend pages

+++ backend/controllers/scheduleController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Schedule = require('../models/scheduleModel')
+const User = require('../models/userModel')
 
 // @desc    Get Schedules
 // @route   GET /api/schedules
 // @access  Private
 const getSchedules = asyncHandler(async (req, res) => {
-    const schedules = await Schedule.find()
+    const schedules = await Schedule.find({ user: req.user.id })
 
     res.status(200).json(schedules)
 })
@@ -27,6 +28,7 @@ const createSchedule = asyncHandler(async (req, res) => {
         day: req.body.day,
         startHour: req.body.startHour,
         endHour: req.body.endHour,
+        user: req.user.id
     })
     res.status(200).json(schedule)
 })
@@ -43,6 +45,18 @@ const updateSchedule = asyncHandler(async (req, res) => {
     if(!schedule) {
         res.status(400)
         throw new Error('Schedule bulunamadı')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user) {
+        res.status(401)
+        throw new Error('Kullanıcı bulunamadı')
+    }
+
+    if(schedule.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('Kullanıcı girişi yapılmadı')
     }
 
     const updatedSchedule = await Schedule.findByIdAndUpdate(req.params.id, req.body, { new: true, })
@@ -62,6 +76,18 @@ const deleteSchedule = asyncHandler(async (req, res) => {
     if(!schedule) {
         res.status(400)
         throw new Error('Schedule bulunamadı')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user) {
+        res.status(401)
+        throw new Error('Kullanıcı bulunamadı')
+    }
+
+    if(schedule.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('Kullanıcı girişi yapılmadı')
     }
 
     await Schedule.findByIdAndDelete(req.params.id)
